feat: New trim edit page

+++ app/Http/Controllers/VideoController.php
@@ -68,23 +68,25 @@ class VideoController extends Controller
     }
 
     public function update(Request $request, $tag){
-        $video = Video::byTag($tag);
-        $ffmpeg = \FFMpeg\FFMpeg::create([
-            'ffmpeg.binaries'  => 'C:/FFmpeg/bin/ffmpeg.exe',
-            'ffprobe.binaries' => 'C:/FFmpeg/bin/ffprobe.exe',
-        ]);
-        $ff = $ffmpeg->open(public_path('storage/videos/'.$video->file));
+        if($request->filled('start', 'end')) {
+            $video = Video::byTag($tag);
+            $ffmpeg = \FFMpeg\FFMpeg::create([
+                'ffmpeg.binaries' => 'C:/FFmpeg/bin/ffmpeg.exe',
+                'ffprobe.binaries' => 'C:/FFmpeg/bin/ffprobe.exe',
+            ]);
+            $ff = $ffmpeg->open(public_path('storage/videos/' . $video->file));
 
-        $ff->filters()->clip(\FFMpeg\Coordinate\TimeCode::fromSeconds($request->start), \FFMpeg\Coordinate\TimeCode::fromSeconds($request->end));
+            $ff->filters()->clip(\FFMpeg\Coordinate\TimeCode::fromSeconds($request->start), \FFMpeg\Coordinate\TimeCode::fromSeconds($request->end));
 
-        $file_new = preg_match('/'.$video->tag.'-(.*).mp4/', $video->file, $output);
+            $file_new = preg_match('/' . $video->tag . '-(.*).mp4/', $video->file, $output);
 
-        $file_new = $video->tag.'-'.time().$output[1].'.mp4';
-        $path_new = 'storage/videos/'.$file_new;
-        $ff->save(new \FFMpeg\Format\Video\X264(), public_path($path_new));
-        Storage::delete('public/videos/'.$video->file);
-        $video->file = $file_new;
-        $video->save();
+            $file_new = $video->tag . '-' . time() . $output[1] . '.mp4';
+            $path_new = 'storage/videos/' . $file_new;
+            $ff->save(new \FFMpeg\Format\Video\X264(), public_path($path_new));
+            Storage::delete('public/videos/' . $video->file);
+            $video->file = $file_new;
+            $video->save();
+        }
     }
 
     public function destroy($tag)
